adding 2step list feature
A list file could be named '.2step.list' or '.list'.
A 2step list used in two steps to get a value to replace the list
placeholder.

In the first step it will be executed without parameters
and it should return a list for selection. Items (lines) in this
list could be a lot more descriptive, they could be rows from a table.

In the second step the script will be executed with the selected by
the user line. The script should parse this line and return a string
which will replace the placeholder.